+++ crates/nu-protocol/src/dataframe/mod.rs
@@ -1,8 +1,8 @@
 pub mod nu_dataframe;
-pub mod nu_grouptuples;
+pub mod nu_groupby;
 
 pub use nu_dataframe::NuDataFrame;
-pub use nu_grouptuples::NuGroupBy;
+pub use nu_groupby::NuGroupBy;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, PartialEq, PartialOrd, Eq, Ord, Hash, Serialize, Deserialize)]
